+++ jkit-sql/src/main/java/com/alianga/jkit/sql/ast/SqlStatementType.java
@@ -23,5 +23,6 @@ public enum SqlStatementType {
     SHOW,
     CALL,
     GRANT,
+    REVOKE,
     OTHER
 }